refactor cluster commands
